+++ CHANGELOG.md
@@ -8,6 +8,10 @@
   [Samuel Giddins](https://github.com/segiddins)
   [#176](https://github.com/CocoaPods/Core/issues/176)
 
+* [Set] Sources are used in the order in which they are provided
+  [Thomas Visser](https://github.com/Thomvis)
+  [#2556](https://github.com/CocoaPods/CocoaPods/issues/2556)
+
 ## 0.34.1
 
 ##### Bug Fixes